+++ greencoffee/src/main/java/com/mauriciotogneri/greencoffee/GreenCoffeeTest.java
@@ -4,11 +4,7 @@ import com.mauriciotogneri.greencoffee.exceptions.DuplicatedStepDefinitionExcept
 import com.mauriciotogneri.greencoffee.exceptions.NoStepsDefinedException;
 import com.mauriciotogneri.greencoffee.exceptions.StepDefinitionNotFoundException;
 
-import java.io.File;
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Locale;
@@ -71,15 +67,10 @@ public class GreenCoffeeTest
         }
         catch (Exception e)
         {
-            if (scenarioConfig.screenshotOnFail())
+            if (scenarioConfig.hasScreenshotProvider())
             {
-                DateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss", Locale.getDefault());
-                String fileName = String.format("%s - %s.jpg", scenario.name(), dateFormat.format(new Date()));
-                File screenshotFile = new File(getTargetContext().getExternalFilesDir("screenshots"), fileName);
-
-
                 ScreenshotProvider screenshotProvider = scenarioConfig.screenshotProvider();
-                screenshotProvider.takeScreenshot(screenshotFile);
+                screenshotProvider.takeScreenshot(scenario);
             }
 
             throw e;